data/requirements.txt
torch
transformers
pinecone-client
python-dotenv
requests
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils
psycopg2-binary
redis
alembic